form_app/requirements.txt
Django>=3.2,<4.0
djangorestframework
django-allauth
gunicorn
psycopg2-binary
djangorestframework-simplejwt
django-cors-headers
python-dotenv
